Feat: replace api calls and add trivial moves on dialogs and drawers part of SWM-153 related to #52
