+++ src/app/(after-login)/(top)/workbookDetail/[workbookId]/_components/workbook-information.tsx
@@ -3,7 +3,7 @@
 import { useParams } from 'next/navigation';
 import WorkbookEditButtons from '@/app/(after-login)/(top)/workbookDetail/[workbookId]/_components/workbook-edit-buttons';
 import { useQuery } from '@tanstack/react-query';
-import { readWorkbookById } from '@/app/(after-login)/(top)/workbookDetail/workbook-detail-api';
+import { readWorkbookById } from '@/app/(after-login)/(top)/workbookDetail/[workbookId]/workbook-detail-api';
 import WorkbookCard from '@/app/(after-login)/(top)/workbookDetail/[workbookId]/_components/workbook-card';
 
 const WorkbookInformation = () => {
